README.md
Don’t miss out on this exclusive opportunity to earn up to 4800 USDT in bonuses and enjoy a 50% discount fee on your trades. Whether you’re in Germany, the USA, or anywhere in the world, the Phemex platform combined with the HMXFT5 invitation code offers a powerful start to your crypto journey in 2025.